make generic page mount function
